+++ src/main/java/com/agenticcp/core/domain/cloud/adapter/outbound/CredentialProviderPortAdapter.java
@@ -7,7 +7,7 @@ import com.agenticcp.core.domain.cloud.exception.CloudErrorCode;
 import com.agenticcp.core.domain.cloud.port.model.CloudSessionCredential;
 import com.agenticcp.core.domain.cloud.port.outbound.CredentialProviderPort;
 import com.agenticcp.core.domain.cloud.repository.CloudAccountRepository;
-import com.agenticcp.core.domain.cloud.service.AwsCredentialManager;
+import com.agenticcp.core.domain.cloud.adapter.outbound.aws.AwsCredentialManager;
 import com.agenticcp.core.domain.cloud.service.SessionCacheService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
