+++ src/main/java/com/agenticcp/core/domain/cloud/service/CloudAccountUseCaseService.java
@@ -3,10 +3,7 @@ package com.agenticcp.core.domain.cloud.service;
 import com.agenticcp.core.common.context.TenantContextHolder;
 import com.agenticcp.core.common.enums.CommonErrorCode;
 import com.agenticcp.core.common.exception.BusinessException;
-import com.agenticcp.core.domain.cloud.command.DeleteCredentialCommand;
-import com.agenticcp.core.domain.cloud.command.ResolveCredentialCommand;
-import com.agenticcp.core.domain.cloud.command.StoreCredentialCommand;
-import com.agenticcp.core.domain.cloud.dto.*;
+import com.agenticcp.core.domain.cloud.port.model.account.*;
 import com.agenticcp.core.domain.cloud.entity.CloudAccount;
 import com.agenticcp.core.domain.cloud.entity.CloudAccountCredential;
 import com.agenticcp.core.domain.cloud.entity.CloudProvider;
@@ -20,7 +17,7 @@ import com.agenticcp.core.domain.cloud.port.outbound.AccountValidationPort;
 import com.agenticcp.core.domain.cloud.port.outbound.AuditEventPort;
 import com.agenticcp.core.domain.cloud.port.outbound.CredentialProviderPort;
 import com.agenticcp.core.domain.cloud.repository.CloudAccountCredentialRepository;
-import com.agenticcp.core.domain.cloud.service.AwsCredentialManager;
+import com.agenticcp.core.domain.cloud.adapter.outbound.aws.AwsCredentialManager;
 import com.agenticcp.core.domain.cloud.repository.CloudAccountRepository;
 import com.agenticcp.core.domain.cloud.repository.CloudProviderRepository;
 import com.agenticcp.core.domain.tenant.entity.Tenant;
